move code in utils
